server_monitor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncssh==2.14.0
pyyaml==6.0.1
jinja2==3.1.2
aiofiles==23.2.1
websockets==12.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
numpy>=1.24.3
pandas>=2.0.3
pydantic==2.5.0
aiohttp==3.9.0
setuptools>=65.0.0
bcrypt==4.0.1
pyjwt==2.8.0
python-multipart==0.0.6